+++ source/plugins/language.visual.wyliozero/visual/code_wyliolab.js
@@ -404,7 +404,7 @@ module.exports = function (blockly) {
 		var statements_function = Blockly.Python.statementToCode(block, 'function');
 		var functionName = Blockly.Python.uniqueName('function');
 		// TODO: Assemble Python into code variable.
-		var code =  '\ndef ' + functionName + '():\n\t' + statements_function.toString()
+		var code =  '\ndef ' + functionName + '():\n' + statements_function.toString()
 				+ value_pin + '.when_held = ' + functionName + '\n\n';
 		return code;
 	};
@@ -415,7 +415,7 @@ module.exports = function (blockly) {
 		var statements_function = Blockly.Python.statementToCode(block, 'function');
 		var functionName = Blockly.Python.uniqueName('function');
 		// TODO: Assemble Python into code variable.
-		var code =  '\ndef ' + functionName + '():\n\t' + statements_function.toString()
+		var code =  '\ndef ' + functionName + '():\n' + statements_function.toString()
 				+ value_pin + '.when_pressed = ' + functionName + '\n\n';
 		return code;
 	};