+++ source/plugins/language.visual.gpiozero/visual/code_gpiozero.js
@@ -182,10 +182,10 @@ module.exports = function (blockly) {
 	Blockly.Python['gpiozero_button_when_held'] = function(block) {
 		Blockly.Python.gpiozero_setUp();
 		var value_pin = Blockly.Python.valueToCode(block, 'pin', Blockly.Python.ORDER_ATOMIC);
-		var statements_function = Blockly.Python.statementToCode(block, 'function');
+		var statements_function = Blockly.Python.statementToCode(block, 'function') || '  pass\n\n';
 		var functionName = Blockly.Python.gpiozero_uniqueName('function');
 		// TODO: Assemble Python into code variable.
-		var code =  '\ndef ' + functionName + '():\n\t' + statements_function.toString()
+		var code =  '\ndef ' + functionName + '():\n' + statements_function.toString()
 				+ value_pin + '.when_held = ' + functionName + '\n\n';
 		return code;
 	};
@@ -193,7 +193,7 @@ module.exports = function (blockly) {
 	Blockly.Python['gpiozero_button_when_pressed'] = function(block) {
 		Blockly.Python.gpiozero_setUp();
 		var value_pin = Blockly.Python.valueToCode(block, 'pin', Blockly.Python.ORDER_ATOMIC);
-		var statements_function = Blockly.Python.statementToCode(block, 'function');
+		var statements_function = Blockly.Python.statementToCode(block, 'function') || '  pass\n\n';
 		var functionName = Blockly.Python.gpiozero_uniqueName('function');
 		// TODO: Assemble Python into code variable.
 		var code =  '\ndef ' + functionName + '():\n' + statements_function.toString()
@@ -204,7 +204,7 @@ module.exports = function (blockly) {
 	Blockly.Python['gpiozero_button_when_released'] = function(block) {
 		Blockly.Python.gpiozero_setUp();
 		var value_pin = Blockly.Python.valueToCode(block, 'pin', Blockly.Python.ORDER_ATOMIC);
-		var statements_function = Blockly.Python.statementToCode(block, 'function');
+		var statements_function = Blockly.Python.statementToCode(block, 'function') || '  pass\n\n';
 		var functionName = Blockly.Python.gpiozero_uniqueName('function');
 		// TODO: Assemble Python into code variable.
 		var code =  '\ndef ' + functionName + '():\n' + statements_function.toString()
@@ -368,10 +368,10 @@ module.exports = function (blockly) {
 	Blockly.Python['gpiozero_light_when_dark'] = function(block) {
 		Blockly.Python.gpiozero_setUp();
 		var value_pin = Blockly.Python.valueToCode(block, 'pin', Blockly.Python.ORDER_ATOMIC);
-		var statements_function = Blockly.Python.statementToCode(block, 'function');
+		var statements_function = Blockly.Python.statementToCode(block, 'function') || '  pass\n\n';
 		var functionName = Blockly.Python.gpiozero_uniqueName('function');
 		// TODO: Assemble Python into code variable.
-		var code =  '\ndef ' + functionName + '():\n\t' + statements_function.toString()
+		var code =  '\ndef ' + functionName + '():\n' + statements_function.toString()
 				+ value_pin + '.when_dark = ' + functionName + '\n\n';
 		return code;
 	};
@@ -379,10 +379,10 @@ module.exports = function (blockly) {
 	Blockly.Python['gpiozero_light_when_light'] = function(block) {
 		Blockly.Python.gpiozero_setUp();
 		var value_pin = Blockly.Python.valueToCode(block, 'pin', Blockly.Python.ORDER_ATOMIC);
-		var statements_function = Blockly.Python.statementToCode(block, 'function');
+		var statements_function = Blockly.Python.statementToCode(block, 'function') || '  pass\n\n';
 		var functionName = Blockly.Python.gpiozero_uniqueName('function');
 		// TODO: Assemble Python into code variable.
-		var code =  '\ndef ' + functionName + '():\n\t' + statements_function.toString()
+		var code =  '\ndef ' + functionName + '():\n' + statements_function.toString()
 				+ value_pin + '.when_light = ' + functionName + '\n\n';
 		return code;
 	};