+++ webpack.browser.config.js
@@ -146,6 +146,10 @@ module.exports = env => {
 				{
 					test: /\.txt$/i,
 					use: 'raw-loader',
+				},
+				{
+					test: /\.ttf$/,
+					use: ['file-loader']
 				}
 				// {
 				// 	test: /\.ts$/,