+++ source/plugins/projects.editor.ace/index.js
@@ -3,7 +3,7 @@ import Ace from './views/AceEditor.vue';
 export default function setup (options, imports, register)
 {
 	const studio = imports;
-	studio.projects.registerEditor('EDITOR_ACE',[''], Ace);
+	studio.projects.registerEditor('EDITOR_ACE',[/makefile(.*)?/i], Ace);
 	
 	register (null, {});
 }